+++ src/Domain/Repository/HappeningParticipationRepositoryInterface.php
@@ -154,4 +154,6 @@ interface HappeningParticipationRepositoryInterface
     public function findBySpeaker(User $user, Event $event): array;
 
     public function hasHappeningParticipant(Event $event): bool;
+
+    public function getPreviousMandatoryEvaluation(Event $event, User $user, \DateTimeInterface $begin): ?HappeningParticipation;
 }